refactor(#38): exception 패키지 경로 수정
