+++ Back/src/main/java/com/example/back/api/auth/service/AuthService.java
@@ -3,9 +3,9 @@ package com.example.back.api.auth.service;
 import com.example.back.api.auth.controller.request.AuthRequest;
 import com.example.back.api.auth.service.jwt.JwtTokenProvider;
 import com.example.back.api.auth.service.response.AuthResponse;
-import com.example.back.common.exception.ExceptionMessage;
-import com.example.back.common.exception.MemberException;
-import com.example.back.common.exception.TokenException;
+import com.example.back.global.exception.ExceptionMessage;
+import com.example.back.global.exception.MemberException;
+import com.example.back.global.exception.TokenException;
 import com.example.back.domain.auth.member.Member;
 import com.example.back.domain.auth.member.repository.MemberRepository;
 import com.example.back.domain.token.jwt.JwtToken;
