+++ Back/src/main/java/com/example/back/global/exception/MemberException.java
@@ -1,4 +1,4 @@
-package com.example.back.common.exception;
+package com.example.back.global.exception;
 
 public class MemberException extends BlueRoseException {
     public MemberException(ExceptionMessage message) {
